requirements.txt
Jinja2==3.0.1
aiofiles==0.7.0
fastapi==0.68.1
google-cloud-bigquery==2.28.1
gunicorn==20.1.0
uvicorn==0.15.0